requirements.txt
fastapi
sqlmodels
uvicorn